cover

+++ tools/patman/database.py
@@ -8,10 +8,11 @@
 import os
 import sqlite3
 
+from u_boot_pylib import tools
 from u_boot_pylib import tout
 
-# Schema version
-LATEST = 2
+# Schema version (version 0 means there is no database yet)
+LATEST = 3
 
 
 class Database:
@@ -24,18 +25,16 @@ class Database:
     def start(self):
         """Open the database read for use"""
         self.open_it()
-        self.ensure_exists()
         self.migrate_to(LATEST)
 
     def open_it(self):
-        """Open the database read for use"""
+        """Open the database ready for use, creating it if necessary"""
         if not os.path.exists(self.db_path):
             tout.warning(f'Creating new database {self.db_path}')
         self.con = sqlite3.connect(self.db_path, autocommit=False)
         self.cur = self.con.cursor()
 
-    def create_v0(self):
-        tout.info('Create database v0')
+    def create_v1(self):
         self.cur.execute(
             'CREATE TABLE series (id INTEGER PRIMARY KEY AUTOINCREMENT,'
             'name UNIQUE, desc, archived BIT)')
@@ -63,14 +62,14 @@ class Database:
             'CREATE TABLE settings (name UNIQUE, proj_id INT, link_name)')
         self.con.commit()
 
-    def _migrate_to_v1(self):
+    def _migrate_to_v2(self):
         """Add a schema_version table"""
         self.cur.execute('CREATE TABLE schema_version (version INTEGER)')
 
-    def _migrate_to_v2(self):
+    def _migrate_to_v3(self):
         """Store the number of cover-letter comments in the schema"""
-        self.cur.execute('ALTER TABLE series ADD COLUMN link')
-        self.cur.execute('ALTER TABLE series ADD COLUMN cover_num_comments')
+        self.cur.execute('ALTER TABLE ser_ver ADD COLUMN cover_id')
+        self.cur.execute('ALTER TABLE ser_ver ADD COLUMN cover_num_comments')
 
     def migrate_to(self, dest_version):
         while True:
@@ -78,33 +77,42 @@ class Database:
             if version == dest_version:
                 break
 
+            self.close()
+            tools.write_file(f'{self.db_path}old.v{version}',
+                             tools.read_file(self.db_path))
+
             version += 1
             tout.info(f'Update database to v{version}')
+            self.open_it()
             if version == 1:
-                self._migrate_to_v1()
+                self.create_v1()
             elif version == 2:
                 self._migrate_to_v2()
-            self.cur.execute('DELETE FROM schema_version')
-            self.cur.execute('INSERT INTO schema_version (version) VALUES (?)',
-                             (version,))
+            elif version == 3:
+                self._migrate_to_v3()
+
+            # Save the new version if we have a schema_version table
+            if version > 1:
+                self.cur.execute('DELETE FROM schema_version')
+                self.cur.execute('INSERT INTO schema_version (version) VALUES (?)',
+                                 (version,))
             self.con.commit()
 
     def get_schema_version(self):
+        # If there is no database at all, assume v0
         version = 0
-        try:
-            # If there is no schema, assume v0
-            self.cur.execute('SELECT version FROM schema_version')
-            version = self.cur.fetchone()[0]
-        except sqlite3.OperationalError:
-            version = 0
-        return version
-
-    def ensure_exists(self):
-        # If there is no database at all, create one
         try:
             self.cur.execute('SELECT name FROM series')
         except sqlite3.OperationalError:
-            self.create_v0()
+            return 0
+
+        # If there is no schema, assume v1
+        try:
+            self.cur.execute('SELECT version FROM schema_version')
+            version = self.cur.fetchone()[0]
+        except sqlite3.OperationalError:
+            return 1
+        return version
 
     def commit(self):
         self.con.commit()
